requirements.txt
# API clients
openai>=0.27.0
anthropic>=0.4.0

# Data processing & analysis
pandas>=1.5.0
numpy>=1.22.0

# Visualization
matplotlib>=3.5.0

# Web app
flask>=2.2.0
python-dotenv>=1.0.0

# Evaluation metrics
sacrebleu>=2.2.0
nltk>=3.8.0
comet>=1.0.2    # COMET MT evaluation
bleurt>=0.0.6   # BLEURT MT evaluation
